pyproject.toml
[project]
name = "restaurant-review-project"
version = "0.1.0"
description = ""
authors = [
    {name = "somy",email = "fuureelf@naver.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pymysql (>=1.1.1,<2.0.0)",
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-cleanup (>=9.0.0,<10.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)"
]
packages = [
    { include = "config" },
    { include = "users" },
    { include = "restaurants" },
    { include = "reviews" }
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
